Add build constraints for other unsupported OS
