+++ crypto/pkcs11/pkcs11helpers_test.go
@@ -1,3 +1,5 @@
+// +build cgo
+
 /*
    Copyright The ocicrypt Authors.
 
@@ -13,6 +15,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
+
 
 package pkcs11
